Added Prototype 0 Chip-8 Emulator
The Prototype 0 Chip-8 Emulator is intended to outline how the emulator will organize its member variables, member functions, and classes in later Prototypes. Currently its only output is to output a short sound when run. (In later prototypes this sound will be output when the loaded Chip-8 program instructs it to be)